разбивает код на модули
