+++ js/utils/check-max-length.js
@@ -1,0 +1,5 @@
+function checkMaxLength(checkString, maxLength) {
+  return (maxLength >= checkString.length);
+}
+
+export {checkMaxLength};
