+++ js/utils/get-random-elements.js
@@ -1,0 +1,7 @@
+import {getRandomNumber} from './get-random-number.js';
+
+function getRandomElements(elements) {
+  return elements[getRandomNumber(0, elements.length-1)];
+}
+
+export {getRandomElements};
